Feat: Modularized profile page with reusable components
- Separated profile functionalities into structured components for better maintainability.
- Implemented , , , and .
- Added a dedicated  for handling image uploads and validation.
- Integrated Redux for user profile management and updates.
- Ensured a scalable and maintainable architecture with centralized imports ().

This improves code readability, reusability, and makes profile management more structured.
